pyproject.toml
[project]
name = "locness-datawriter"
version = "0.1.0"
description = "DynamoDB to Google Drive CSV data writer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
]
